requirements.txt
alembic==1.6.5
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dataclass-factory==2.11
Faker==8.5.1
greenlet==1.1.0
idna==2.10
importlib-metadata==4.5.0
logging-tree==1.9
Mako==1.1.4
MarkupSafe==2.0.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
requests==2.25.1
requests-toolbelt==0.9.1
rope==0.19.0
six==1.16.0
SQLAlchemy==1.4.17
text-unidecode==1.3
typing-extensions==3.10.0.0
urllib3==1.26.5
zipp==3.4.1
